Fix implicit conversions errors

+++ source/d2d/drawable/ishape.d
@@ -6,5 +6,5 @@ import d2d.drawable.vertex;
 * IShape
 */
 interface IShape {
-	public void draw(ref Vertex[] verts, ref uint[] indices);
+	public void draw(ref Vertex[] verts, ref ulong[] indices);
 }
